requirements.txt
# requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
diffusers==0.19.3
accelerate==0.20.3
openai==0.27.8
pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
opencv-python==4.8.0.74
scipy==1.10.1
streamlit==1.24.0
gradio==3.39.0